botforge/requirements.txt
python-telegram-bot[job-queue]>=20.7
psutil>=5.9
python-dotenv>=1.0
httpx>=0.27
aiofiles>=23.0
APScheduler>=3.10
